Add fixes and tests for engine idle timeout handling
Updates test driver to handle empty frames incoming and fixes several
issues with the engine idle handling.  Tests ported from proton-j to
cover some of the known cases of issues with tick.

+++ qpid-proton4j-engine/src/main/java/org/apache/qpid/proton4j/engine/impl/ProtonFrameEncodingHandler.java
@@ -69,7 +69,7 @@ public class ProtonFrameEncodingHandler implements EngineHandler {
 
     @Override
     public void handleWrite(EngineHandlerContext context, SaslPerformative performative) {
-        context.fireWrite(writeFrame(saslEncoder, saslEncoderState, performative, null, SASL_FRAME_TYPE, (short) 0, configuration.getOutboundMaxFrameSize(), null));
+        context.fireWrite(writeFrame(saslEncoder, saslEncoderState, performative, null, SASL_FRAME_TYPE, 0, configuration.getOutboundMaxFrameSize(), null));
     }
 
     private ProtonBuffer writeFrame(Encoder encoder, EncoderState encoderState, Object performative, ProtonBuffer payload, byte frameType, int channel, int maxFrameSize, Runnable onPayloadTooLarge) {
